move timestamp/isolationLevel to minigu-common and remove transaction/txnManager trait
