+++ minigu/common/src/timestamp.rs
@@ -7,8 +7,23 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, OnceLock};
 
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 
-use crate::error::TimestampError;
+/// Errors that can occur during timestamp and transaction ID operations.
+#[derive(Error, Debug)]
+pub enum TimestampError {
+    #[error("expected commit-ts, but got txn-id ({0})")]
+    WrongDomainCommit(u64),
+
+    #[error("expected txn-id, but got commit-ts ({0})")]
+    WrongDomainTxnId(u64),
+
+    #[error("commit-ts overflow, reached {0}")]
+    CommitTsOverflow(u64),
+
+    #[error("txn-id overflow, reached {0}")]
+    TxnIdOverflow(u64),
+}
 
 /// Represents a commit timestamp used for multi-version concurrency control (MVCC).
 /// It can either represent a transaction ID which starts from 1 << 63,
