+++ minigu/storage/tests/common/mod.rs
@@ -9,7 +9,7 @@ use minigu_storage::model::vertex::Vertex;
 use minigu_storage::tp::MemoryGraph;
 use minigu_storage::tp::checkpoint::CheckpointManagerConfig;
 use minigu_storage::wal::graph_wal::WalManagerConfig;
-use minigu_transaction::{GraphTxnManager, IsolationLevel, Transaction};
+use minigu_transaction::IsolationLevel;
 
 pub const PERSON_LABEL_ID: LabelId = LabelId::new(1).unwrap();
 pub const FRIEND_LABEL_ID: LabelId = LabelId::new(1).unwrap();
